requirements.txt
aiogram==3.1.1
SQLAlchemy==2.0.23
aiosqlite==0.19.0
WalletPay==1.3.1
APScheduler==3.10.4
binance-pay-connector==1.0.0
qrcode==7.4.2
